+++ CHANGELOG.md
@@ -1,13 +1,29 @@
+## [1.0.294](https://github.com/alternun-development/alternun/compare/v1.0.294-dev.0...v1.0.294) (2026-06-29)
+
+### Bug Fixes
+
+- **repo:** docs: sync root README for v1.0.294
+
+
+
+
+
+
+## [1.0.294](https://github.com/alternun-development/alternun/compare/v1.0.293...v1.0.294) (2026-06-29)
+
+### Bug Fixes
+
+- **repo:** fix(repo,api,mobile): .gitignore, airs.repository, AppInfoFooter, Dashboard
+
+### Bug Fixes
+
+- **repo,api,mobile:** .gitignore, airs.repository, AppInfoFooter, Dashboard ([7836867](https://github.com/alternun-development/alternun/commit/7836867037b7b0b3f306621d3952418dfda4a500))
+
 ## [1.0.293](https://github.com/alternun-development/alternun/compare/v1.0.293-dev.0...v1.0.293) (2026-06-28)
 
 ### Bug Fixes
 
 - **repo:** docs: sync root README for v1.0.293
-
-
-
-
-
 
 ## [1.0.293](https://github.com/alternun-development/alternun/compare/v1.0.292...v1.0.293) (2026-06-28)